extract psm tags input into separate component
