+++ app/templates/components/gh-post-settings-menu.hbs
@@ -65,20 +65,7 @@
                     }}
                 </div>
 
-                <div class="form-group">
-                    <label for="tag-input">Tags</label>
-                    {{gh-selectize
-                        id="tag-input"
-                        multiple=true
-                        selection=model.tags
-                        content=availableTags
-                        optionValuePath="content.uuid"
-                        optionLabelPath="content.name"
-                        openOnFocus=false
-                        create-item="addTag"
-                        remove-item="removeTag"
-                        plugins="remove_button, drag_drop"}}
-                </div>
+                {{gh-psm-tags-input post=model}}
 
                 {{#gh-form-group errors=model.errors hasValidated=model.hasValidated property="customExcerpt"}}
                     <label for="custom-excerpt">Excerpt</label>
